+++ .gitignore
@@ -163,3 +163,4 @@ cython_debug/
 .ruff_cache
 
 *.db
+*.html
